Pass ANIMA_APP instead of ANIMA_SUBPACKAGE
The difference is the 'app' is the name of the executable,
which on Windows includes the ".exe" suffix.

Use .removesuffix(".exe") to take care of fixing up
the places that shouldn't be.

The Windows packages weren't packaging their .exes because
their package_data was using the .exe-less form and therefore
wasn't matching any files. This should fix that.